Fixed merge issues and added relative urls to v2 api docs
