+++ app/assets/javascripts/lara-typescript.js
