+++ docs/lara-plugin-api-V2/docs/interfaces/ieventlistener.md
@@ -23,7 +23,7 @@
 
 **● listener**: *`function`*
 
-*Defined in [api/decorate-content.ts:5](../../../lara-plugin-api-V2/src/api/decorate-content.ts#L5)*
+*Defined in [api/decorate-content.ts:5](https://github.com/concord-consortium/lara/blob/5741bf58/lara-plugin-api-V2/src/api/decorate-content.ts#L5)*
 
 #### Type declaration
 ▸(evt: *`Event`*): `void`
@@ -43,7 +43,7 @@ ___
 
 **● type**: *`string`*
 
-*Defined in [api/decorate-content.ts:4](../../../lara-plugin-api-V2/src/api/decorate-content.ts#L4)*
+*Defined in [api/decorate-content.ts:4](https://github.com/concord-consortium/lara/blob/5741bf58/lara-plugin-api-V2/src/api/decorate-content.ts#L4)*
 
 ___
