+++ docs/lara-plugin-api-V2/docs/interfaces/iplugin.md
@@ -1,0 +1,12 @@
+[LARA Plugin API](../README.md) > [IPlugin](../interfaces/iplugin.md)
+
+# Interface: IPlugin
+
+## Hierarchy
+
+**IPlugin**
+
+## Index
+
+---
+
