+++ docs/lara-plugin-api-V2/docs/interfaces/iruntimecontextexperimentalfeatures.md
@@ -22,7 +22,7 @@
 
 **● clickToPlayId**: *`string` \| `null`*
 
-*Defined in [api/plugins.ts:62](../../../lara-plugin-api-V2/src/api/plugins.ts#L62)*
+*Defined in [api/plugins.ts:62](https://github.com/concord-consortium/lara/blob/5741bf58/lara-plugin-api-V2/src/api/plugins.ts#L62)*
 
 DOM id of click to play overlay if enabled.
