+++ docs/lara-plugin-api/interfaces/iclassinfo.md
@@ -27,23 +27,7 @@
 
 **● class_hash**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:140](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L140)*
-=======
-*Defined in [types.ts:140](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L140)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:140](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L140)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:140](lara-typescript/src/plugin-api/types.ts#L140)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:140](../../../lara-typescript/src/plugin-api/types.ts#L140)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="id"></a>
@@ -52,23 +36,7 @@ ___
 
 **● id**: *`number`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:138](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L138)*
-=======
-*Defined in [types.ts:138](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L138)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:138](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L138)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:138](lara-typescript/src/plugin-api/types.ts#L138)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:138](../../../lara-typescript/src/plugin-api/types.ts#L138)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="offerings"></a>
@@ -77,23 +45,7 @@ ___
 
 **● offerings**: *[IOffering](ioffering.md)[]*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:143](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L143)*
-=======
-*Defined in [types.ts:143](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L143)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:143](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L143)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:143](lara-typescript/src/plugin-api/types.ts#L143)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:143](../../../lara-typescript/src/plugin-api/types.ts#L143)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="students"></a>
@@ -102,23 +54,7 @@ ___
 
 **● students**: *[IUser](iuser.md)[]*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:142](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L142)*
-=======
-*Defined in [types.ts:142](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L142)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:142](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L142)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:142](lara-typescript/src/plugin-api/types.ts#L142)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:142](../../../lara-typescript/src/plugin-api/types.ts#L142)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="teachers"></a>
@@ -127,23 +63,7 @@ ___
 
 **● teachers**: *[IUser](iuser.md)[]*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:141](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L141)*
-=======
-*Defined in [types.ts:141](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L141)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:141](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L141)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:141](lara-typescript/src/plugin-api/types.ts#L141)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:141](../../../lara-typescript/src/plugin-api/types.ts#L141)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="uri"></a>
@@ -152,23 +72,7 @@ ___
 
 **● uri**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:139](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L139)*
-=======
-*Defined in [types.ts:139](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L139)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:139](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L139)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:139](lara-typescript/src/plugin-api/types.ts#L139)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:139](../../../lara-typescript/src/plugin-api/types.ts#L139)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
