+++ docs/lara-plugin-api/interfaces/ieventlistener.md
@@ -23,23 +23,7 @@
 
 **● listener**: *`function`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [decorate-content.ts:5](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/decorate-content.ts#L5)*
-=======
-*Defined in [decorate-content.ts:5](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/decorate-content.ts#L5)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [decorate-content.ts:5](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/decorate-content.ts#L5)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [decorate-content.ts:5](lara-typescript/src/plugin-api/decorate-content.ts#L5)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [decorate-content.ts:5](../../../lara-typescript/src/plugin-api/decorate-content.ts#L5)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 #### Type declaration
 ▸(evt: *`Event`*): `void`
@@ -59,23 +43,7 @@ ___
 
 **● type**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [decorate-content.ts:4](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/decorate-content.ts#L4)*
-=======
-*Defined in [decorate-content.ts:4](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/decorate-content.ts#L4)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [decorate-content.ts:4](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/decorate-content.ts#L4)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [decorate-content.ts:4](lara-typescript/src/plugin-api/decorate-content.ts#L4)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [decorate-content.ts:4](../../../lara-typescript/src/plugin-api/decorate-content.ts#L4)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
