+++ docs/lara-plugin-api/interfaces/iinteractivestate.md
@@ -27,23 +27,7 @@
 
 **● activity_name**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:152](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L152)*
-=======
-*Defined in [types.ts:152](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L152)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:152](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L152)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:152](lara-typescript/src/plugin-api/types.ts#L152)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:152](../../../lara-typescript/src/plugin-api/types.ts#L152)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="id"></a>
@@ -52,23 +36,7 @@ ___
 
 **● id**: *`number`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:147](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L147)*
-=======
-*Defined in [types.ts:147](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L147)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:147](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L147)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:147](lara-typescript/src/plugin-api/types.ts#L147)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:147](../../../lara-typescript/src/plugin-api/types.ts#L147)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="interactive_name"></a>
@@ -77,23 +45,7 @@ ___
 
 **● interactive_name**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:150](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L150)*
-=======
-*Defined in [types.ts:150](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L150)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:150](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L150)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:150](lara-typescript/src/plugin-api/types.ts#L150)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:150](../../../lara-typescript/src/plugin-api/types.ts#L150)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="interactive_state_url"></a>
@@ -102,23 +54,7 @@ ___
 
 **● interactive_state_url**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:151](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L151)*
-=======
-*Defined in [types.ts:151](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L151)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:151](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L151)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:151](lara-typescript/src/plugin-api/types.ts#L151)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:151](../../../lara-typescript/src/plugin-api/types.ts#L151)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="key"></a>
@@ -127,23 +63,7 @@ ___
 
 **● key**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:148](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L148)*
-=======
-*Defined in [types.ts:148](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L148)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:148](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L148)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:148](lara-typescript/src/plugin-api/types.ts#L148)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:148](../../../lara-typescript/src/plugin-api/types.ts#L148)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="raw_data"></a>
@@ -152,23 +72,7 @@ ___
 
 **● raw_data**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:149](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L149)*
-=======
-*Defined in [types.ts:149](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L149)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:149](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L149)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:149](lara-typescript/src/plugin-api/types.ts#L149)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:149](../../../lara-typescript/src/plugin-api/types.ts#L149)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
