+++ docs/lara-plugin-api/interfaces/ijwtclaims.md
@@ -26,23 +26,7 @@
 
 **● claims**: *[IPortalClaims](iportalclaims.md)*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:117](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L117)*
-=======
-*Defined in [types.ts:117](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L117)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:117](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L117)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:117](lara-typescript/src/plugin-api/types.ts#L117)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:117](../../../lara-typescript/src/plugin-api/types.ts#L117)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="class_info_url"></a>
@@ -51,23 +35,7 @@ ___
 
 **● class_info_url**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:116](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L116)*
-=======
-*Defined in [types.ts:116](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L116)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:116](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L116)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:116](lara-typescript/src/plugin-api/types.ts#L116)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:116](../../../lara-typescript/src/plugin-api/types.ts#L116)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="domain"></a>
@@ -76,23 +44,7 @@ ___
 
 **● domain**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:113](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L113)*
-=======
-*Defined in [types.ts:113](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L113)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:113](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L113)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:113](lara-typescript/src/plugin-api/types.ts#L113)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:113](../../../lara-typescript/src/plugin-api/types.ts#L113)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="externalid"></a>
@@ -101,23 +53,7 @@ ___
 
 **● externalId**: *`number`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:115](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L115)*
-=======
-*Defined in [types.ts:115](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L115)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:115](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L115)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:115](lara-typescript/src/plugin-api/types.ts#L115)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:115](../../../lara-typescript/src/plugin-api/types.ts#L115)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="returnurl"></a>
@@ -126,23 +62,7 @@ ___
 
 **● returnUrl**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:114](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L114)*
-=======
-*Defined in [types.ts:114](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L114)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:114](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L114)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:114](lara-typescript/src/plugin-api/types.ts#L114)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:114](../../../lara-typescript/src/plugin-api/types.ts#L114)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
