+++ docs/lara-plugin-api/interfaces/ijwtresponse.md
@@ -23,23 +23,7 @@
 
 **● claims**: *[IJwtClaims](ijwtclaims.md) \| `__type`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:122](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L122)*
-=======
-*Defined in [types.ts:122](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L122)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:122](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L122)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:122](lara-typescript/src/plugin-api/types.ts#L122)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:122](../../../lara-typescript/src/plugin-api/types.ts#L122)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="token"></a>
@@ -48,23 +32,7 @@ ___
 
 **● token**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:121](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L121)*
-=======
-*Defined in [types.ts:121](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L121)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:121](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L121)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:121](lara-typescript/src/plugin-api/types.ts#L121)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:121](../../../lara-typescript/src/plugin-api/types.ts#L121)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
