+++ docs/lara-plugin-api/interfaces/ilogdata.md
@@ -26,23 +26,7 @@ That's the minimal set of properties that needs to be provided. All the other pr
 
 **● event**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:160](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L160)*
-=======
-*Defined in [types.ts:160](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L160)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:160](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L160)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:160](lara-typescript/src/plugin-api/types.ts#L160)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:160](../../../lara-typescript/src/plugin-api/types.ts#L160)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="event_value"></a>
@@ -51,23 +35,7 @@ ___
 
 **● event_value**: *`any`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:161](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L161)*
-=======
-*Defined in [types.ts:161](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L161)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:161](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L161)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:161](lara-typescript/src/plugin-api/types.ts#L161)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:161](../../../lara-typescript/src/plugin-api/types.ts#L161)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="parameters"></a>
@@ -76,23 +44,7 @@ ___
 
 **● parameters**: *`any`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:162](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L162)*
-=======
-*Defined in [types.ts:162](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L162)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:162](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L162)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:162](lara-typescript/src/plugin-api/types.ts#L162)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:162](../../../lara-typescript/src/plugin-api/types.ts#L162)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
