+++ docs/lara-plugin-api/interfaces/ioffering.md
@@ -24,23 +24,7 @@
 
 **● id**: *`number`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:132](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L132)*
-=======
-*Defined in [types.ts:132](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L132)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:132](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L132)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:132](lara-typescript/src/plugin-api/types.ts#L132)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:132](../../../lara-typescript/src/plugin-api/types.ts#L132)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="name"></a>
@@ -49,23 +33,7 @@ ___
 
 **● name**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:133](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L133)*
-=======
-*Defined in [types.ts:133](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L133)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:133](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L133)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:133](lara-typescript/src/plugin-api/types.ts#L133)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:133](../../../lara-typescript/src/plugin-api/types.ts#L133)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="url"></a>
@@ -74,23 +42,7 @@ ___
 
 **● url**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:134](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L134)*
-=======
-*Defined in [types.ts:134](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L134)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:134](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L134)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:134](lara-typescript/src/plugin-api/types.ts#L134)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:134](../../../lara-typescript/src/plugin-api/types.ts#L134)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
