+++ docs/lara-plugin-api/interfaces/iportalclaims.md
@@ -25,23 +25,7 @@
 
 **● class_hash**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:108](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L108)*
-=======
-*Defined in [types.ts:108](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L108)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:108](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L108)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:108](lara-typescript/src/plugin-api/types.ts#L108)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:108](../../../lara-typescript/src/plugin-api/types.ts#L108)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="offering_id"></a>
@@ -50,23 +34,7 @@ ___
 
 **● offering_id**: *`number`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:109](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L109)*
-=======
-*Defined in [types.ts:109](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L109)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:109](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L109)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:109](lara-typescript/src/plugin-api/types.ts#L109)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:109](../../../lara-typescript/src/plugin-api/types.ts#L109)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="user_id"></a>
@@ -75,23 +43,7 @@ ___
 
 **● user_id**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:107](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L107)*
-=======
-*Defined in [types.ts:107](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L107)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:107](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L107)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:107](lara-typescript/src/plugin-api/types.ts#L107)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:107](../../../lara-typescript/src/plugin-api/types.ts#L107)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="user_type"></a>
@@ -100,23 +52,7 @@ ___
 
 **● user_type**: *"learner" \| "teacher"*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:106](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L106)*
-=======
-*Defined in [types.ts:106](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L106)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:106](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L106)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:106](lara-typescript/src/plugin-api/types.ts#L106)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:106](../../../lara-typescript/src/plugin-api/types.ts#L106)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
