+++ docs/lara-plugin-api/interfaces/iuser.md
@@ -24,23 +24,7 @@
 
 **● first_name**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:127](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L127)*
-=======
-*Defined in [types.ts:127](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L127)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:127](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L127)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:127](lara-typescript/src/plugin-api/types.ts#L127)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:127](../../../lara-typescript/src/plugin-api/types.ts#L127)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="id"></a>
@@ -49,23 +33,7 @@ ___
 
 **● id**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:126](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L126)*
-=======
-*Defined in [types.ts:126](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L126)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:126](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L126)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:126](lara-typescript/src/plugin-api/types.ts#L126)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:126](../../../lara-typescript/src/plugin-api/types.ts#L126)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
 <a id="last_name"></a>
@@ -74,23 +42,7 @@ ___
 
 **● last_name**: *`string`*
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-<<<<<<< HEAD
-*Defined in [types.ts:128](https://github.com/concord-consortium/lara/blob/7771e1f1/lara-typescript/src/plugin-api/types.ts#L128)*
-=======
-*Defined in [types.ts:128](https://github.com/concord-consortium/lara/blob/5ed958f8/lara-typescript/src/plugin-api/types.ts#L128)*
->>>>>>> Added v3 plugin label override with internal label [#166193888]The plugin system now ignores the label provided by the plugin when itregisters and instead uses an internally generated label set before eachplugin is loaded.  This allows multiple copies of the same plugin to existon a page.
-=======
-*Defined in [types.ts:128](https://github.com/concord-consortium/lara/blob/master/lara-typescript/src/plugin-api/types.ts#L128)*
->>>>>>> Added lara-typescript api doc build step to rewrite blob to master in docs
-=======
-*Defined in [types.ts:128](lara-typescript/src/plugin-api/types.ts#L128)*
->>>>>>> Updated api doc replacement to use npm packages and relative paths
-=======
 *Defined in [types.ts:128](../../../lara-typescript/src/plugin-api/types.ts#L128)*
->>>>>>> Change fix docs paths to use script to give correct relative paths
 
 ___
