+++ lara-plugin-api-V2/fix-doc-paths.js
@@ -1,0 +1,5 @@
+#! /usr/bin/env node
+
+const replace = require("../lara-typescript/replace")
+
+replace(`${__dirname}/docs/`, "../")
