+++ lara-plugin-api-V2/package.json
@@ -11,7 +11,10 @@
   "scripts": {
     "build": "npm-run-all lint:build clean build:webpack build:docs",
     "build:webpack": "webpack --devtool false && cp ./dist/lara-plugin-api.js ../app/assets/javascripts/lara-plugin-api-V2.js",
-    "build:docs": "typedoc --theme markdown --readme DOC-INTRO.md --ignoreCompilerErrors --name 'LARA Plugin API' --excludeNotExported --excludeExternals --excludePrivate --mode file --out ./docs ./src/lara-plugin-api.ts ./src/api/* && cp -R ./docs ../docs/lara-plugin-api-V2/",
+    "build:docs": "npm-run-all build:docs:plugin-api build:docs:fix-doc-paths build:docs:copy-files",
+    "build:docs:plugin-api": "typedoc --theme markdown --readme DOC-INTRO.md --ignoreCompilerErrors --name 'LARA Plugin API' --excludeNotExported --excludeExternals --excludePrivate --mode file --out ./docs ./src/lara-plugin-api.ts ./src/api/* && cp -R ./docs ../docs/lara-plugin-api-V2/",
+    "build:docs:fix-doc-paths": "./fix-doc-paths.js",
+    "build:docs:copy-files": "cpx './docs/**/*.*' ../docs/lara-plugin-api-V2/",
     "clean": "rimraf dist docs ../docs/lara-plugin-api-V2",
     "lint": "tslint -c tslint.json 'src/**/*.{ts,tsx}'",
     "lint:build": "tslint -c tslint-build.json 'src/**/*.{ts,tsx}'",
@@ -62,6 +65,7 @@
     "@types/jqueryui": "^1.12.7",
     "@types/react": "^16.8.13",
     "babel-core": "^6.26.3",
+    "cpx": "^1.5.0",
     "fork-ts-checker-webpack-plugin": "^1.0.0",
     "jest": "^23.6.0",
     "jquery": "^3.3.1",
